Update save-transaction-transact-sql.md
I think that the check for XACT_STATE() = 1 is much clearer than XACT_STATE() <> -1
As whilst I understand that XACT_STATE() returning 0 would not be expected in this scenario so there would be little functional consequence, but a state of 0 would also be a case for not rolling back, just as a return of -1 would be.  I think therefore this makes it easier for others to understand.